xcute: Delete 'manager'
